Updated obs-bounce with new option
I really liked this script but I found it a bit annoying how I couldn't change the timer until next throw, and was forced to go into the file itself and edit it. So I took the time and added some stuff and there is now two new options that people can use. I've tested it with several different combinations of settings and it all seems to work. 

- "Shortest time until next throw (seconds)" 
-"Longest amount of time until next throw(seconds)" 
Range is 1 to 3600 seconds. Meaning that the longest time until next throw is an hour. Setting the two values will obviously not randomize the throws anymore and will instead just countdown the time set and throw the little cute cat or whatever the source is. 

NOTE: I have never edited a OBS script, nor Lua before so if any of the code I added is not usable for some reason that I'm not aware of, please tell me because I would love to learn more about it. I didn't even know Lua existed before today.  Same goes for github, I'm honestly so lost in this process so please do tell me If I've misunderstood anything